Added a check to the domain commands for extra positional arguments. (#6658)
What changed?
Added a check to error out when positional arguments are specified in domain commands.

Why?
In this pr: #6552 we fixed the behaviour of the --cl c1 c2 argument in the domain commands to be instead --cl c1,c2. This was needed due to the update of the urfave CLI framework to v2 (see the PR for more info).

In much of our documentation we still mention that the clusters in the domain commands should be specified as --cl c1 c2 however as mention this no longer works.

Before this PR, anything after the first positional argument was ignored, so the command cadence --do foo domain register --cl c1 c2 --st securityToken was interpreted as cadence --do foo domain register --cl c1  with c2 --st securityToken as ignored positional arguments. This command will error out with "No permission to do this operation." which is very misleading.

The new error will make the user immediately aware of the problem and direct them to the help documentation.

How did you test it?

Potential risks

Release notes

Documentation Changes
Yes wee should fix documentation, at least in these two places:

https://cadenceworkflow.io/docs/cli
https://cadenceworkflow.io/docs/concepts/cross-dc-replication