Added exponential retry to the domain cache (#6676)
What changed?
Added a proper retry policy to the domain cache so it does not just retry every second

Why?
The load from retrying every second can cause too much load on the database.

How did you test it?
Unit test

Potential risks
Might cause domain data to be out of date, but we are already in that situation when it's failing.

Release notes

Documentation Changes