Restructure Admin TaskList commands to operate on multiple types (#6712)
While Activity/Decision TaskLists with a given name are independent, they typically represent a common purpose and we have never operated on just a single type. Handling them entirely separately within the admin commands doubles the amount of commands needed to run to operate on the TaskList and makes it harder to understand whether a given TaskList name is used for decision and/or activity tasks. Particularly as we move TaskList partition data from dynamic config to the DB it's important that we provide good operator ergonomics to avoid errors.

Update the Admin TaskList commands with the following:
- The absence of the `tasklisttype` flag now indicates both Activity and Decision. It previously was inconsistent between defaulting to Decision or being required.
- The describe command will output one row per type specified rather than a single row.
- The describe command will output that a TaskList has 1 read/write partition rather than no output if there is no config.
- The describe command will no longer output the list of pollers and instead print out the number of pollers. This functionality is still available via the non-admin version of the command, which does nothing other than printing the pollers.
- The describe command now supports json as an output format.
- The list command will consolidate Activity and Decision TaskLists with a given name to a single row, including separate columns for the number of decision and activity pollers.
- The list command now sorts its output by TaskList name.
- The list command now supports json as an output format.
- The update-partition command now operates on multiple TaskListTypes at once rather than Decision or Activity. It performs a safety check on both before performing any updates.