Bugfix/fixing history deletion on reset failure (1/2) - adding documentation/tests (#6740)
I was looking for a data 'leak' (or data that isn't being cleaned up correctly). I didn't find the one I was looking for, but ended up finding another. This documents (and will, subsequently), cleanup a problem with reset-workflows leaving garbage data around.

The major interesting point of this PR is in the unit-test for common/persistence/nosql/nosql_history_store_test.go. I'm not making any behaviour changes, I'll fix / unskip the tests for fixing the todos in a subsequent PR.