Wrap matching peer provider (#6715)
What changed?

Added a new dynamic config to control the mode of the shard distributor.
The hashrings are now initialized in the server, so they can be passed to the Multiresolver as well as the shard distribtor.
The initialization of the shard distributor client is moved to the server as well, as the shard distributor client should be passed to the Multiresolver, which is created on the server.
The shard distributor now uses the raw hashing instead of using the MultiResolver.
A wrapper around the hashring for matching decides whether to use the shard distributor or to use the hash ring to find a shard owner.
Why?
This will allow us to start having the shard distributor decide the owners of the different shards. This is part of the migration to the new shard distributor.
Additionally, we will start having metrics on things like latencies, etc.

How did you test it?
Tested locally, and via unit tests

Potential risks
This touches the memebership module, which is very core to Cadence. The new feature is behind a feature flag so there should be no impact, but touching core modules is risky.

Release notes

Documentation Changes