Adjust task isolation metrics (#6784)
Remove the isolation group that we're currently dispatching the task as. It isn't useful. We only care whether the original matches the poller.

Add an explicit metric to count when the original group matches the poller. This is a lot easier to monitor and query compared to finding where the tags match.