Bugfix/fixing history deletion on reset (2/2) (#6742)
What changed?

This (hopefully) addresses two issues around workflow history cleanup:

That branched workflows are not cleaning up the parent fully (as is the case with a simple reset workflow, where the ancestor branch's earlier nodes will simply remain).
That - while not observed in the wild - there seems to me to be a reasonable chance that branches being cleaned up before their ancestors will end up trimming the history_nodes of their parents if the deletion doesn't occur in the creation order (and there's a couple of ways in which that can occur). I don't have immediate proof of this, but I think there's sufficient risk that given consecutive failovers and divergent history, that a shorter lived branch might break a more long-lived ancestor (which, admittedly, is very much an edge-case, but nevertheless).
As described in #6740, I was trying to fix some problems where we're seeing data get not cleaned up correctly in several cases. Reset is clearly one (that I was able to reproduce locally). As per the earlier unit-test, I wasn't super sure what the existing behaviour was, so spent some time ensuring I could cover it well before making this change.