Document timer task page tokens' risky serialization needs (#6821)
Discovered due to a changed test in https://github.com/cadence-workflow/cadence/pull/6800
So I am writing this down so an accidental cleanup doesn't break task replication.

This _can probably_ be changed at some point, and it'd probably be worth it, but it'd need to be done in two or three fully-deployed-and-all-data-updated stages (support both, create in new format, remove old support) and I don't feel like doing that right now.