Remove duplicate buffered activity events (#6966)
There appears to be a very rare recurring issue from persistence where buffered events are duplicated. While we likely know the cause, we're unable to generally reproduce it. We can reliably detect when these duplicates occur though, so remove them.

If these events are written to the history it causes issues in the client, resulting in the workers reporting non-determinism.