Increase Activity Rate Limiter Responsiveness (#7079)
The TaskList rate limiter allows for controlling the rate at which Cadence dispatches tasks to Workers. It is applied only to Activities, with each PollForActivityTaskRequest including the desired rate limit. The rate limit isn't persisted with the TaskListInfo, which means it needs to be re-applied any time TaskList ownership changes.

Introduce a new Limiter implementation specific for TaskLists that eagerly adopts lower reported rate limits, and introduce a new metric for the limit set on a TaskList to provide additional visibility into it.