Better lifecycle management in the shard executor library (#7095)
What changed?

Better shutdown of the executor process loop
Made the updateShardAssignment function use waitgroups, so it blocks until all the processors has started or stopped
Made sure we do not stop heartbeating if an assignment takes a long time.
If we are still running an old assignment when we get a new one, we will skip the new one.
Protected the state of the managed processor with an atomic int
Why?
Make sure the life cycle management of the executor library is robust

How did you test it?
Unit tests

Potential risks

Release notes

Documentation Changes